# public/api-docs/api-docs.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "CRM RESTful API Docs v1.0",
        "description": "Just about CRM RESTful API Documentation. Enjoyed `Created by H@$b1`",
        "contact": {
            "email": "justhasby@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://157.245.55.185:8088/api/v1"
        }
    ],
...
    "paths": {
...
        "/pertanyaan": {
            "post": {
                "tags": [
                    "Pertanyaan-Controller"
                ],
                "summary": "getAllPertanyaanPaginate",
                "operationId": "getAllPertanyaanPaginate",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/getIndexPaginate"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/200"
                    },
                    "400": {
                        "$ref": "#/components/responses/400"
                    },
                    "401": {
                        "$ref": "#/components/responses/401"
                    },
                    "403": {
                        "$ref": "#/components/responses/403"
                    },
                    "404": {
                        "$ref": "#/components/responses/404"
                    },
                    "500": {
                        "$ref": "#/components/responses/500"
                    }
                }
            },
            "get": {
                "tags": [
                    "Pertanyaan-Controller"
                ],
                "summary": "getPertanyaanById",
                "operationId": "getPertanyaanById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "schema": {
                            "$ref": "#/components/parameters/getById"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/200"
                    },
                    "400": {
                        "$ref": "#/components/responses/400"
                    },
                    "401": {
                        "$ref": "#/components/responses/401"
                    },
                    "403": {
                        "$ref": "#/components/responses/403"
                    },
                    "404": {
                        "$ref": "#/components/responses/404"
                    },
                    "500": {
                        "$ref": "#/components/responses/500"
                    }
                }
            },
            "put": {
                "tags": [
                    "Pertanyaan-Controller"
                ],
                "summary": "updatePertanyaan",
                "operationId": "updatePertanyaan",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Pertanyaan"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/200"
                    },
                    "400": {
                        "$ref": "#/components/responses/400"
                    },
                    "401": {
                        "$ref": "#/components/responses/401"
                    },
                    "403": {
                        "$ref": "#/components/responses/403"
                    },
                    "404": {
                        "$ref": "#/components/responses/404"
                    },
                    "500": {
                        "$ref": "#/components/responses/500"
                    }
                }
            }
        },
...
    },
    "components": {
        "responses": {
            "400": {
                "description": "The Bad Request."
            },
            "401": {
                "description": "Unauthorized."
            },
            "403": {
                "description": "Forbidden."
            },
            "404": {
                "description": "The specified resource was not found."
            },
            "200": {
                "description": "Success."
            },
            "500": {
                "description": "Inter Server Error."
            }
        },
        "parameters": {
...
            "getById": {
                "name": "id",
                "in": "query"
            },
...
        },
        "schemas": {
            "getIndexPaginate": {
                "type": "object",
                "required": [
                    "page",
                    "size",
                    "sortField",
                    "sortOrder"
                ],
                "properties": {
                    "page": {
                        "type": "integer",
                        "example": 1
                    },
                    "size": {
                        "type": "integer",
                        "example": 5
                    },
                    "sortField": {
                        "type": "string",
                        "example": "created_at"
                    },
                    "sortOrder": {
                        "type": "string",
                        "example": "desc"
                    },
                    "filters": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Filter"
                        }
                    }
                },
                "xml": {
                    "name": "getIndexPaginate"
                }
            },
            "Filter": {
                "type": "object",
                "required": [
                    "key",
                    "operation",
                    "value"
                ],
                "properties": {
                    "key": {
                        "type": "string",
                        "example": null
                    },
                    "operation": {
                        "type": "string",
                        "example": null
                    },
                    "value": {
                        "type": "string",
                        "default": null,
                        "example": null
                    }
                },
                "xml": {
                    "name": "Filter"
                }
            },
...
            "Pertanyaan": {
                "type": "object",
                "properties": {
                    "id_pertanyaan": {
                        "type": "integer",
                        "description": "Input id_pertanyaan"
                    },
                    "id_instansi": {
                        "type": "integer",
                        "description": "Input id_instansi -> lookup from instansi"
                    },
                    "id_tema": {
                        "type": "integer",
                        "description": "Input id_tema -> lookup from tema"
                    },
                    "pertanyaan": {
                        "type": "string",
                        "description": "Input Pertanyaan"
                    }
                },
                "xml": {
                    "name": "Pertanyaan"
                }
            },
...
        },
        "securitySchemes": {
            "CRM_Auth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
    "security": [
        {
            "CRM_Auth": []
        }
    ]
}